Fixes to pass VSCode Intellisence checks
Often, the attribute((section())) statement was used with an extra
set of parentheses around the section name. This caused issues for
VSCode and resulted in the files being marked as having errors, making
it harder to spot actual errors.

This commit removes the extra parentheses. It also adds a #define for
'''__builtin_preserve_access_index()''' if VSCODE has been defined. This
allows developers to define VSCODE in their VSCODE C/C++ settings in
order to stop Intellisense complaining about the _(P) macro which is
used in probe_read() calls.

Signed-off-by: Kevin Sheldrake <kevin.sheldrake@isovalent.com>